pyproject.toml
[tool.poetry]
name = "chatline"
version = "0.1.0"
authors = ["Baze <basileaw@gmail.com>"]
description = "A chat interface implementation"
packages = [{ include = "chatline" }]
readme = "README.md"
repository = "https://github.com/anotherbazeinthewall/chatline-interface"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
boto3 = "*"
httpx = "*"
rich = "*"
prompt-toolkit = "*"

[tool.poetry.group.dev.dependencies]
fastapi = "*"
uvicorn = "*"
poethepoet = "*"
python-dotenv = "*"

[tool.poe.tasks]
embedded-client = "python examples/client.py"
embedded-client-log = "python examples/client.py --enable-logging --log-file logs/chatline_debug.log"
serve = "python examples/server.py"
remote-client = "python examples/client.py -e http://127.0.0.1:8000/chat"
remote-client-log = "python examples/client.py -e http://127.0.0.1:8000/chat --enable-logging --log-file logs/chatline_debug.log"
